myapp/requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
Flask==2.3.2
Flask-Session==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.4
